refactor: Исправить названия пакетов, документация WorkLocationScreen
